BREAKING: Refactors file validation
